# openapi.yaml
openapi: 3.0.3
info:
  title: MindWeaver API
  description: |
    MindWeaver 个人知识/研究动态监控工具 API

    功能特性：
    - RSS/Atom 订阅源管理
    - 条目获取、去重、存储和检索
    - 分类管理
    - 过滤规则管理
    - 自动调度获取
    - AI 摘要生成
    - 邮件摘要推送

  version: 1.0.0
  contact:
    name: MindWeaver
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
paths:
...
  /entry/{entry_id}:
    get:
      tags: [Pages]
      summary: 条目详情页面
      description: 返回条目详情页面 (HTML)
      parameters:
        - name: entry_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: HTML 页面
        '404':
          description: 条目不存在

  # ==================== Feeds API ====================
  /api/feeds:
    get:
      tags: [Feeds]
      summary: 获取所有订阅源
      description: 列出所有 RSS/Atom 订阅源
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedListResponse'

    post:
      tags: [Feeds]
      summary: 创建订阅源
      description: 添加新的 RSS/Atom 订阅源
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/categories:
    get:
      tags: [Categories]
      summary: 获取所有分类
      parameters:
        - name: enabled_only
          in: query
          schema:
            type: boolean
            default: false
          description: 仅返回启用的分类
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Category'

    post:
      tags: [Categories]
      summary: 创建分类
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Category'
                  message:
                    type: string
...
  /api/filter-rules:
    get:
      tags: [Filter Rules]
      summary: 获取所有过滤规则
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FilterRule'

    post:
      tags: [Filter Rules]
      summary: 创建过滤规则
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRuleCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/FilterRule'
                  message:
                    type: string
...
  /api/entries/{id}:
    get:
      tags: [Entries]
      summary: 获取单个条目
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Entry'

    put:
      tags: [Entries]
      summary: 更新条目
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntryUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Entry'
                  message:
                    type: string

    delete:
      tags: [Entries]
      summary: 删除条目
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 操作成功
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: 错误信息
        status:
          type: integer
          example: 400
    FeedListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Feed'
    FeedResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Feed'
        message:
          type: string

    # ==================== Feed 模型 ====================
    Feed:
      type: object
      properties:
        id:
          type: integer
          description: 订阅源 ID
        url:
          type: string
          format: uri
          description: RSS/Atom 订阅源 URL
        name:
          type: string
          description: 订阅源名称
        description:
          type: string
          description: 订阅源描述
        enabled:
          type: boolean
          description: 是否启用
          default: true
        fetch_interval_minutes:
          type: integer
          description: 获取间隔（分钟）
          default: 60
        max_entries_per_fetch:
          type: integer
          description: 每次获取最大条目数
          default: 50
        fetch_only_recent:
          type: boolean
          description: 仅获取最近条目
          default: false
        created_at:
          type: string
          format: date-time
          description: 创建时间
        updated_at:
          type: string
          format: date-time
          description: 更新时间
        last_fetched_at:
          type: string
          format: date-time
          description: 上次获取时间
        fetch_error_count:
          type: integer
          description: 连续错误次数
          default: 0
        last_error:
          type: string
          description: 上次错误信息
        last_error_at:
          type: string
          format: date-time
          description: 上次错误时间
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
          description: 关联的分类列表
    FeedCreate:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          description: RSS/Atom 订阅源 URL
        name:
          type: string
          description: 订阅源名称
        description:
          type: string
          description: 订阅源描述
        enabled:
          type: boolean
          default: true
        fetch_interval_minutes:
          type: integer
          default: 60
        max_entries_per_fetch:
          type: integer
          default: 50
        fetch_only_recent:
          type: boolean
          default: false
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          description: 分类名称
        description:
          type: string
          description: 分类描述
        color:
          type: string
          description: 分类颜色（Hex 格式，如 #FF5733）
        icon:
          type: string
          description: 分类图标
        enabled:
          type: boolean
          default: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        feed_count:
          type: integer
          description: 关联的订阅源数量
    CategoryCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        description:
          type: string
          maxLength: 200
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          example: '#FF5733'
        icon:
          type: string
        enabled:
          type: boolean
          default: true
...
    FilterRule:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          description: 规则名称
        enabled:
          type: boolean
          default: true
        rule_type:
          type: string
          enum: [keyword, regex, tag, language]
          description: 规则类型
        match_type:
          type: string
          enum: [include, exclude]
          description: 匹配类型
        pattern:
          type: string
          description: 匹配模式
        priority:
          type: integer
          description: 优先级（数字越小优先级越高）
          default: 10
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FilterRuleCreate:
      type: object
      required:
        - name
        - rule_type
        - match_type
        - pattern
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 50
        rule_type:
          type: string
          enum: [keyword, regex, tag, language]
        match_type:
          type: string
          enum: [include, exclude]
        pattern:
          type: string
          minLength: 1
        priority:
          type: integer
          default: 10
        enabled:
          type: boolean
          default: true
...
    Entry:
      type: object
      properties:
        id:
          type: integer
        feed_id:
          type: integer
          description: 所属订阅源 ID
        title:
          type: string
          description: 条目标题
        link:
          type: string
          format: uri
          description: 原始链接
        author:
          type: string
          description: 作者
        summary:
          type: string
          description: AI 生成的摘要
        content:
          type: string
          description: 完整内容
        published_at:
          type: string
          format: date-time
          description: 发布时间
        fetched_at:
          type: string
          format: date-time
          description: 获取时间
        tags:
          type: array
          items:
            type: string
          description: 标签列表
        language:
          type: string
          description: 语言代码
        reading_time_seconds:
          type: integer
          description: 预计阅读时间（秒）
        title_hash:
          type: string
          description: 标题哈希（用于去重）
        link_hash:
          type: string
          description: 链接哈希（用于去重）
        content_hash:
          type: string
          description: 内容哈希（用于去重）
        enabled:
          type: boolean
          description: 是否启用（未读=true，已读=false）
          default: true
    EntryUpdate:
      type: object
      properties:
        title:
          type: string
        summary:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        enabled:
          type: boolean
          description: 启用/禁用（对应未读/已读）

    # ==================== Digest Log 模型 ====================
...